// dsa-client/src/app/enums/special-category-data.enum.ts
export enum SpecialCategoryData {
  NOT_SPECIAL_CATEGORY_DATA = 'NOT_SPECIAL_CATEGORY_DATA',
  RACIAL = 'RACIAL',
  POLITICAL = 'POLITICAL',
  RELIGIOUS = 'RELIGIOUS',
  TRADE_UNION = 'TRADE_UNION',
  GENETIC = 'GENETIC',
  BIOMETRIC_ID = 'BIOMETRIC_ID',
  HEALTH = 'HEALTH',
  SEX_LIFE = 'SEX_LIFE',
  SEXUAL_ORIENT = 'SEXUAL_ORIENT'
}