// dsa-client/src/app/enums/article-9-condition.enum.ts
export enum Article9Condition {
  NOT_APPLICABLE = 'NOT_APPLICABLE',
  EXPLICIT_CONSENT = 'EXPLICIT_CONSENT',
  EMPLOYMENT = 'EMPLOYMENT',
  VITAL_INTERESTS = 'VITAL_INTERESTS',
  NOT_FOR_PROFIT ='NOT_FOR_PROFIT',
  DATA_MADE_PUBLIC = 'DATA_MADE_PUBLIC',
  LEGAL = 'LEGAL',
  REASONS_OF_PUBLIC_INTEREST = 'REASONS_OF_PUBLIC_INTEREST',
  HEALTH_OR_OCCUP_CARE = 'HEALTH_OR_OCCUP_CARE',
  PUBLIC_HEALTH = 'PUBLIC_HEALTH',
  ARCHIVING = 'ARCHIVING'
}