// dsa-client/src/app/enums/data-content-type.enum.ts
export enum DataContentType {
  NOT_SPECIFIED = "NOT_SPECIFIED",
  STRUCTURED_ELECTRONIC_DATA = "STRUCTURED_ELECTRONIC_DATA",
  DATABASE_RECORD = "DATABASE_RECORD",
  UNSTRUCTURED_ELECTRONIC_DATA = "UNSTRUCTURED_ELECTRONIC_DATA",
  ELECTRONIC_DOCUMENT = "ELECTRONIC_DOCUMENT",
  IMAGE_OR_BIOMETRIC_FILE = "IMAGE_OR_BIOMETRIC_FILE",
  AUDIO_VISUAL_RECORDING = "AUDIO_VISUAL_RECORDING",
  LOG_OR_TELEMETRY_DATA = "LOG_OR_TELEMETRY_DATA",
  PAPER_DOCUMENT = "PAPER_DOCUMENT",
  OTHER_MEDIUM = "OTHER_MEDIUM"
}